Game schemas complete, parameters moved to definition

+++ trunk/open-ig/src/hu/openig/model/BattleInfo.java
@@ -147,7 +147,7 @@ public class BattleInfo {
 	 */
 	public void findHelpers() {
 		World world = attacker.owner.world;
-		final int minDistance = world.params().nearbyDistance();
+		final double minDistance = world.params().nearbyDistance();
 		if (targetFleet != null) {
 			// locate the nearest planet
 			double dmin = Double.MAX_VALUE;
